Methods and accesors for GenomicSignatures object

The default contents of GenomicSignatures object, with a set of getter and setter generic functions, which extract either the assay, colData, or metadata slots of a GenomicSignatures-class object. When you create this object, colData$studies should be populated before adding any information in trainingData slot.

# S4 method for GenomicSignatures
RAVindex(x)

# S4 method for GenomicSignatures
geneSets(x)

# S4 method for GenomicSignatures
updateNote(x)

# S4 method for GenomicSignatures
version(x)

# S4 method for GenomicSignatures
geneSets(x) <- value

# S4 method for GenomicSignatures
updateNote(x) <- value

Arguments

x

A GenomicSignatures object

value

See details.

Value

A GenomicSignatures object for the constructor

Details

assay(x) : RAVindex (= avgLoadings) containing genes x RAVs

metadata(x) : Metadata associated with RAVindex building process

colData(x) : Information on RAVs

Setters

Setter method values (i.e., function(x) <- value):

metadata<- : Assign metadata

coldata<- : Assign extra information associated with RAVs

geneSets<- : A character vector containing the name of gene sets used to annotate average loadings

updateNote<- : A character vector. Describes the main feature of a model construction

Getters

RAVindex : Equivalent to assays(x)$RAVindex

geneSets : Access the metadata(x)$geneSets slot

updateNote : Access the metadata(x)$updateNote slot

version : Access the metadata(x)$version slot

Examples

data(miniRAVmodel)
miniRAVmodel
#> class: PCAGenomicSignatures 
#> dim: 13934 20 
#> metadata(8): cluster size ... version geneSets
#> assays(1): RAVindex
#> rownames(13934): CASKIN1 DDX3Y ... CTC-457E21.9 AC007966.1
#> rowData names(0):
#> colnames(20): RAV1076 RAV338 ... RAV312 RAV468
#> colData names(4): RAV studies silhouetteWidth gsea
#> trainingData(2): PCAsummary MeSH
#> trainingData names(536): DRP000987 SRP059172 ... SRP164913 SRP188526