Validating new dataset

Validating new dataset

.loadingCor(dataset, avgLoading, method = "pearson", scale = FALSE)

Arguments

dataset
A gene expression profile to be validated. Different classes of objects can be used including ExpressionSet, SummarizedExperiment, RangedSummarizedExperiment, or matrix. Rownames (genes) should be in symbol format. If it is a matrix, genes should be in rows and samples in columns.

avgLoading
A matrix with genes by RAVs.

method
A character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.

scale
Default is FALSE. If it is set to TRUE, dataset will be row normalized by rowNorm function.

Value

A matrix of Pearson correlation coefficient (default, defined through method argument) between RAVs (row) and the top 8 PCs from the datasets (column)