PCA on gene expression profile

Performs a principal components analysis on the given data matrix and returns the results as an object of class prcomp.

extractPC(x)

Arguments

x
a numeric or complex matrix (or data frame) which provides the gene expression data for the principal components analysis. Genes in the rows and samples in the columns.

Value

A prcomp object.

See also

Examples

m = matrix(rnorm(100),ncol=5)
extractPC(m)
#> Standard deviations (1, .., p=5):
#> [1] 3.302708e+00 2.438158e+00 2.341146e+00 1.786600e+00 3.214289e-16
#> 
#> Rotation (n x k) = (20 x 5):
#>               PC1         PC2         PC3          PC4         PC5
#>  [1,]  0.15941075  0.21443742  0.11097515 -0.118658049  0.44462676
#>  [2,] -0.07964333 -0.25806106  0.68789939  0.035648797 -0.37938022
#>  [3,] -0.24103561 -0.04251069 -0.03738448 -0.239792689  0.16432569
#>  [4,] -0.27308728  0.05468303 -0.17189738  0.143960908  0.02509908
#>  [5,] -0.21269393  0.33629188 -0.15738918 -0.078719210 -0.42128433
#>  [6,] -0.18414627  0.15268476  0.20914856 -0.004411023 -0.21683073
#>  [7,] -0.20514899 -0.31279927 -0.01962280  0.285156730  0.04034952
#>  [8,] -0.07553045 -0.33383038 -0.24675322  0.248762683  0.11523513
#>  [9,] -0.19991895  0.01496008  0.05413004 -0.051103678  0.02581457
#> [10,]  0.12810801  0.16955408 -0.15852543  0.363389372 -0.19307526
#> [11,]  0.36637335 -0.17178610 -0.16213626 -0.020029425  0.07347223
#> [12,]  0.31416373 -0.33469021  0.09826708  0.235778903 -0.08865577
#> [13,] -0.08926130  0.32990767  0.30876869  0.467460692  0.48829342
#> [14,]  0.31649779  0.19533794  0.27680224 -0.294270685  0.07133737
#> [15,] -0.27491128 -0.05477457 -0.07794636  0.046840183  0.02270141
#> [16,]  0.25025691 -0.19659199 -0.06012648 -0.158682255  0.02549688
#> [17,]  0.16294224  0.21183912 -0.17417224  0.308542686 -0.16628562
#> [18,]  0.22526162  0.27318752  0.14669340  0.314365117 -0.16266181
#> [19,]  0.17912943  0.22124692 -0.23609442 -0.151026054 -0.19110115
#> [20,]  0.25219000 -0.09659323 -0.02229558  0.130509686 -0.06946018