MSigDB GSEA results

MSigDB GSEA results

msigdb_gsea(
  ind,
  RAVmodel,
  category = "C2",
  n = NULL,
  pvalueCutoff = 0.5,
  minGSSize = 10,
  maxGSSize = 500,
  pAdjustMethod = "BH",
  verbose = FALSE,
  seed = FALSE,
  by = "fgsea",
  geneSets = NULL
)

Arguments

ind
An interger. Index of RAV to apply GSEA.

RAVmodel
PCAGenomicSignature object.

category
A character vector representing MSigDB category. Options are "H", "C1", "C2"(default), "C3", "C4", "C5", "C6", and "C7"

n
An interger. The number of top and bottom enriched pathways to plot. Default is NULL and print out all the pathways enriched under pvalueCutoff.

pvalueCutoff
Cutoff for both pvalue and p.adjust. Default is 0.5.

minGSSize
A mininum size of gene set to be analyzed

maxGSSize
A maximum size of gene set to be analyzed

pAdjustMethod
p-value adjustment methods, which will be used as an input for method argument of p.adjust function. Available options are "holm", "hochberg", "hommel", "bonferroni", "BH"(default), "BY", "fdr", "none".

verbose
Logical. Default is FALSE.

seed
Logical. Default is FALSE.

by
Available options are c("fgsea", "DOSE"). Default is "fgsea".

geneSets
Custom genesets to use with MSigDB genesets. It should be in a named list format.

Value

Barplot of GSEA output. Top and bottom n genesets based on NES are plotted and qvalues are denoted by color.