Two-dimentional PCA plot with the PC annotation

Two-dimentional PCA plot with the PC annotation

plotAnnotatedPCA(
  dataset,
  RAVmodel,
  PCnum,
  val_all = NULL,
  scoreCutoff = 0.5,
  nesCutoff = NULL,
  color_by = NULL,
  color_lab = NULL,
  trimed_pathway_len = 45
)

Arguments

dataset

A gene expression profile to be validated. Different classes of objects can be used including ExpressionSet, SummarizedExperiment, RangedSummarizedExperiment, or matrix. Rownames (genes) should be in symbol format. If it is a matrix, genes should be in rows and samples in columns.

RAVmodel

PCAGenomicSignatures-class object

PCnum

A numeric vector length of 2. It should be between 1 and 8.

val_all

The output from validate

scoreCutoff

A numeric value for the minimum correlation. Default 0.5.

nesCutoff

A numeric value for the minimum NES. Default is NULL and the suggested value is 3.

color_by

A named vector with the feature you want to color by. Name should be match with the sample names of the dataset.

color_lab

A name for color legend. If this argument is not provided, the color legend will be labeled as "Color By" by default.

trimed_pathway_len

Positive inter values, which is the display width of pathway names. Default is 45.

Value

Scatter plot and the table with annotation. If enriched pathway didn't pass the scoreCutoff the table will be labeled as "No significant pathways". If any enriched pathway didn't pass the nesCutoff, it will labeled as NA.

Examples

data(miniRAVmodel)
library(bcellViper)
data(bcellViper)
if (FALSE) {
plotAnnotatedPCA(exprs(dset), miniRAVmodel, PCnum = c(1,2))
}