Subset enriched pathways of RAV

Subset enriched pathways of RAV

subsetEnrichedPathways(RAVmodel, ind = NULL, n = 10, both = FALSE)

Arguments

RAVmodel
PCAGenomicSignatures object. Also an output from GSEA can be used.

ind
A numeric vector containing the RAV number you want to check enriched pathways. If not specified, this function returns results from all the RAVs.

n
The number of top and bottom pathways to be selected based on normalized enrichment score (NES).

both
Default is FALSE, where only the top n pathways will be printed. If it is set to TRUE, the output will contain both top and bottom n pathways.

Value

A DataFrame with top and bottom n pathways from the enrichment results.

Examples

data(miniRAVmodel)

# all RAVS in model
subsetEnrichedPathways(miniRAVmodel,n=5)
#> DataFrame with 5 rows and 17 columns
#>                     RAV1076                RAV2538                 RAV338
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_METABOLISM_..    REACTOME_CELL_CYCLE MIPS_PA700_20S_PA28_..
#> Up_2       KEGG_SPLICEOSOME REACTOME_CELL_CYCLE_.. REACTOME_RESPIRATORY..
#> Up_3 REACTOME_ASSEMBLY_OF.. REACTOME_DNA_REPLICA.. REACTOME_RESPIRATORY..
#> Up_4 REACTOME_LATE_PHASE_.. REACTOME_MITOTIC_M_M.. MIPS_55S_RIBOSOME_MI..
#> Up_5 MIPS_39S_RIBOSOMAL_S.. REACTOME_MITOTIC_PRO.. KEGG_PARKINSONS_DISE..
#>                     RAV1467                RAV1139                 RAV884
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_DNA_REPLICA..              DMAP_ERY3              DMAP_ERY3
#> Up_2    REACTOME_CELL_CYCLE REACTOME_CYTOKINE_SI.. REACTOME_METABOLISM_..
#> Up_3 REACTOME_MITOTIC_M_M..     REACTOME_APOPTOSIS KEGG_ALZHEIMERS_DISE..
#> Up_4 REACTOME_CELL_CYCLE_.. KEGG_HUNTINGTONS_DIS.. KEGG_HUNTINGTONS_DIS..
#> Up_5 REACTOME_METABOLISM_.. REACTOME_MRNA_PROCES..        KEGG_CELL_CYCLE
#>                     RAV1614                 RAV294                RAV3071
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_DNA_REPLICA.. REACTOME_MITOTIC_M_M.. KEGG_PURINE_METABOLISM
#> Up_2 REACTOME_MITOTIC_M_M.. REACTOME_CELL_CYCLE_.. MIPS_55S_RIBOSOME_MI..
#> Up_3 REACTOME_CELL_CYCLE_.. REACTOME_DNA_REPLICA..       PID_FOXM1PATHWAY
#> Up_4    REACTOME_CELL_CYCLE REACTOME_CELL_CYCLE_.. REACTOME_E2F_MEDIATE..
#> Up_5 REACTOME_METABOLISM_.. MIPS_55S_RIBOSOME_MI..        PID_ATR_PATHWAY
#>                     RAV1694                 RAV438                 RAV725
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_ADAPTIVE_IM.. REACTOME_METABOLISM_.. REACTOME_ADAPTIVE_IM..
#> Up_2 KEGG_PURINE_METABOLISM              DMAP_ERY5 REACTOME_METABOLISM_..
#> Up_3              DMAP_ERY5              DMAP_ERY4              DMAP_ERY5
#> Up_4 REACTOME_SIGNALING_B.. REACTOME_POST_TRANSL..              DMAP_ERY4
#> Up_5              DMAP_ERY4              DMAP_ERY3              DMAP_ERY3
#>                     RAV1497                 RAV501                 RAV941
#>                 <character>            <character>            <character>
#> Up_1 MIPS_28S_RIBOSOMAL_S.. REACTOME_TRANSCRIPTION IRIS_DendriticCell-C..
#> Up_2 MIPS_55S_RIBOSOME_MI..    REACTOME_DNA_REPAIR       PID_FOXM1PATHWAY
#> Up_3 REACTOME_RESPIRATORY.. REACTOME_METABOLISM_..       PID_PLK1_PATHWAY
#> Up_4 KEGG_OXIDATIVE_PHOSP.. REACTOME_3_UTR_MEDIA.. IRIS_CD4Tcell-Th1-re..
#> Up_5 REACTOME_RESPIRATORY.. REACTOME_METABOLISM_.. REACTOME_SYNTHESIS_O..
#>                      RAV695                RAV1994
#>                 <character>            <character>
#> Up_1 IRIS_Bcell-Memory_Ig..           DMAP_TCELLA6
#> Up_2           DMAP_BCELLA3 SVM T cells CD4 memo..
#> Up_3  IRIS_Bcell-Memory_IgM           DMAP_TCELLA7
#> Up_4       IRIS_Bcell-naive           DMAP_TCELLA8
#> Up_5           DMAP_BCELLA4           DMAP_TCELLA2

# only a specific RAV (note the colnames above)
subsetEnrichedPathways(miniRAVmodel,ind=695,n=5)
#> DataFrame with 5 rows and 1 column
#>                      RAV695
#>                 <character>
#> Up_1 IRIS_Bcell-Memory_Ig..
#> Up_2           DMAP_BCELLA3
#> Up_3  IRIS_Bcell-Memory_IgM
#> Up_4       IRIS_Bcell-naive
#> Up_5           DMAP_BCELLA4