Validation result in data frame

Validation result in data frame

validatedSignatures(
  val_all,
  RAVmodel,
  num.out = 5,
  scoreCutoff = NULL,
  swCutoff = NULL,
  clsizeCutoff = NULL,
  indexOnly = FALSE,
  whichPC = NULL,
  filterMessage = TRUE
)

Arguments

val_all

An output matrix from validate function. If this input is from multiple datasets, only scoreCutoff argument will be considered and other inputs will be ignored.

RAVmodel

PCAGenomicSignatures-class object. RAVmodel used to prepare val_all input.

num.out

A number of highly validated RAVs to output. Default is 5. If any of the cutoff parameters are provided, num.out or the number of filtered RAVs, whichever smaller, will be chosen.

scoreCutoff

A numeric value for the minimum correlation. For multi-studies case, the default is 0.7.

swCutoff

A numeric value for the minimum average silhouette width.

clsizeCutoff

An integer value for the minimum cluster size.

indexOnly

A logical. Under the default (= FALSE), the detailed information on validated RAVs, such as score, average silhouette width, cluster size, is printed. If it is set TRUE, only the RAV number will be printed.

whichPC

An integer value between 1 and 8. PC number of your data to check the validated signatures with. Under the default (NULL), it outputs top scored signatures with any PC of your data.

filterMessage

A logical. Under the default TRUE, any output RAV belong to the filtering list will give a message. Silence this message with filterMessage=FALSE. You can check the filter list using data("filterList").

Value

A subset of the input matrix, which meets the given condition.

Examples

data(miniRAVmodel)
library(bcellViper)
data(bcellViper)
val_all <- validate(dset, miniRAVmodel)
validatedSignatures(val_all, miniRAVmodel, num.out = 3, scoreCutoff = 0)
#> RAV1076 can be filtered based on GSEA_PLIERpriors
#> RAV2538 can be filtered based on GSEA_PLIERpriors
#> RAV338 can be filtered based on GSEA_PLIERpriors
#>             score PC          sw cl_size cl_num
#> RAV1076 0.5950767  2 -0.04447124      10   1076
#> RAV2538 0.5838616  2  0.06996166       4   2538
#> RAV338  0.5709072  2 -0.04683319      21    338