Construct PCAGenomicSignatures object

The default contents of PCAGenomicSignatures object, with a set of accessors and setter generic functions, which extract either the assay, colData, metadata, or trainingData slots of a PCAGenomicSignatures-class object. When you create this object, colData$studies should be populated before adding any information in trainingData slot

PCAGenomicSignatures(...)

Arguments

...
Additional arguments for supporting functions.

Value

PCAGenomicSignatures object with multiple setters or accessors

Details

RAVindex(x) : RAVindex (= avgLoadings) containing genes x RAVs

metadata(x)$cluster : A vector of integers (from 1:k) indicating the cluster to which each point is allocated.

metadata(x)$size : The number of PCs in each cluster.

metadata(x)$k : The number of RAVs.

metadata(x)$n : The number of top PCs from each dataset.

metadata(x)$geneSets : Name of the prior gene sets used to annotate average loadings.

colData(x)$studies : A list of character vectors containing studies contributing to each PC cluster.

colData(x)$silhouetteWidth : A numeric array of average silhouette widths of each clusters

colData(x)$gsea : A list of data frames. Each element is a subset of outputs from clusterProfiler::GSEA function.

Slots

trainingData

A DataFrame class object for metadata associated with training data

Setters

Setter method values (i.e., function(x) <- value):

geneSets<- : A character vector containing the name of gene sets used to annotate average loadings

studies<- : A list of character vectors containing gene sets used to annotate average loadings

gsea<- : A list of data frames. Each element is a subset of output from gseaResult objects.

metadata<- : A list object of metadata

`$<-` : A vector to replace the indicated column in colData

Accessors

All the accessors inherited from SummarizedExperiment are available and the additional accessors for PCAGenomicSignatures specific data are listed below.

RAVindex : Equivalent to the assay(x)

geneSets : Access the metadata(x)$geneSets slot

studies : Access the colData(x)$studies slot

gsea : Access the colData(x)$gsea

`$` : Access a column in colData

trainingData : Access the trainingData slot

mesh : Access the trainingData(x)$MeSH slot

PCAsummary : Access the trainingData(x)$PCAsummary slot

Examples

data(miniRAVmodel)
miniRAVmodel
#> class: PCAGenomicSignatures 
#> dim: 13934 17 
#> metadata(7): cluster size ... updateNote version
#> assays(1): RAVindex
#> rownames(13934): CASKIN1 DDX3Y ... CTC-457E21.9 AC007966.1
#> rowData names(0):
#> colnames(17): RAV1076 RAV2538 ... RAV695 RAV1994
#> colData names(4): RAV studies silhouetteWidth gsea
#> trainingData(2): PCAsummary MeSH
#> trainingData names(536): DRP000987 SRP059172 ... SRP164913 SRP188526