PCA on gene expression profile

Performs a principal components analysis on the given data matrix and returns the results as an object of class prcomp.

extractPC(x)

Arguments

x

a numeric or complex matrix (or data frame) which provides the gene expression data for the principal components analysis. Genes in the rows and samples in the columns.

Value

A prcomp object.

See also

Examples

m = matrix(rnorm(100),ncol=5)
extractPC(m)
#> Standard deviations (1, .., p=5):
#> [1] 2.828616e+00 2.403002e+00 1.788300e+00 1.151328e+00 5.523177e-16
#> 
#> Rotation (n x k) = (20 x 5):
#>               PC1         PC2          PC3          PC4          PC5
#>  [1,] -0.20001451  0.11104080 -0.108361168  0.078416195  0.455782080
#>  [2,]  0.23850218 -0.15275523 -0.030985185  0.407249000 -0.340695544
#>  [3,]  0.17665914 -0.05776792 -0.234004822  0.224423107  0.711061230
#>  [4,]  0.45704542  0.25151281 -0.068731250 -0.247597716  0.129869414
#>  [5,] -0.21876770  0.19902723  0.243148288 -0.341735444  0.084670771
#>  [6,]  0.09911683  0.33916503  0.259060922  0.301006021  0.074975815
#>  [7,] -0.36583658  0.06039575  0.013761678  0.062986795  0.095139771
#>  [8,] -0.34391630  0.16683608  0.004338695 -0.016314606 -0.130258888
#>  [9,] -0.19817194  0.40280492  0.081761147  0.409616678 -0.054165885
#> [10,] -0.14272642 -0.20480371 -0.303429459 -0.130580178 -0.082124868
#> [11,] -0.11826614 -0.03683666  0.137807279 -0.061863483  0.085022692
#> [12,] -0.25531687  0.15366303 -0.369787581  0.058246057 -0.002507256
#> [13,]  0.12715859  0.04055140  0.193608690  0.377813779 -0.006107799
#> [14,]  0.11224405  0.21829763  0.139163871 -0.249094540 -0.040426208
#> [15,] -0.29380290 -0.14707874  0.122078066  0.052841343 -0.025149833
#> [16,] -0.11228416 -0.11571744  0.562118241 -0.108941567  0.138902222
#> [17,]  0.20293366  0.15946499 -0.122100609 -0.203361739 -0.149144998
#> [18,]  0.06728246 -0.34080482  0.203126267  0.202006914 -0.037003220
#> [19,] -0.14461850  0.21700758 -0.308479139  0.104795480 -0.218916037
#> [20,]  0.14169583  0.45766516  0.088467815  0.007359055 -0.069688101