PCA on gene expression profile

Performs a principal components analysis on the given data matrix and returns the results as an object of class prcomp.

extractPC(x)

Arguments

x
a numeric or complex matrix (or data frame) which provides the gene expression data for the principal components analysis. Genes in the rows and samples in the columns.

Value

A prcomp object.

See also

Examples

m = matrix(rnorm(100),ncol=5)
extractPC(m)

#> Standard deviations (1, .., p=5):
#> [1] 3.462628e+00 2.635527e+00 2.048566e+00 1.793355e+00 3.260295e-16
#> 
#> Rotation (n x k) = (20 x 5):
#>               PC1         PC2         PC3         PC4          PC5
#>  [1,] -0.27817111  0.05211119 -0.11572250 -0.13258313  0.446620081
#>  [2,] -0.25781771 -0.10774861  0.19253027 -0.29931773  0.200232784
#>  [3,] -0.18316536 -0.10058578  0.16154914  0.22734894 -0.133518208
#>  [4,] -0.15063325  0.21384779 -0.36145699  0.22607761 -0.474963296
#>  [5,] -0.03920223  0.25439266 -0.42816101 -0.07622255  0.124950051
#>  [6,] -0.17643716  0.05993438  0.07852034  0.10815339 -0.142913545
#>  [7,]  0.04612530 -0.24226732 -0.24885230 -0.25920122  0.192000259
#>  [8,]  0.35603860  0.04776492 -0.15131688 -0.21821680  0.211340008
#>  [9,]  0.33340107  0.03722977 -0.33642680  0.21989622  0.197029951
#> [10,] -0.16287093 -0.43742079 -0.07861413  0.29556024  0.213584920
#> [11,]  0.28832671 -0.22907157  0.38446024  0.10328921 -0.188371288
#> [12,] -0.24629769  0.13943986 -0.06705856  0.01977356 -0.082088228
#> [13,]  0.27247685  0.12347589 -0.01783349 -0.06982633 -0.102853358
#> [14,]  0.07319654 -0.26377437 -0.19423345 -0.31576440 -0.327596488
#> [15,]  0.13534655 -0.41392224 -0.06851865  0.01876046 -0.006876494
#> [16,]  0.11938549 -0.11500067  0.13399079 -0.46023146 -0.165176069
#> [17,]  0.23442431 -0.05144069 -0.17261310 -0.03517784 -0.177602986
#> [18,]  0.10378941 -0.24158519 -0.28505106  0.11795914 -0.113636177
#> [19,] -0.41744609 -0.34635912 -0.26987205 -0.19674922 -0.220256171
#> [20,] -0.06865701  0.29248787  0.04729210 -0.37462158 -0.185998875