PCA on gene expression profile

Performs a principal components analysis on the given data matrix and returns the results as an object of class prcomp.

extractPC(x)

Arguments

x

a numeric or complex matrix (or data frame) which provides the gene expression data for the principal components analysis. Genes in the rows and samples in the columns.

Value

A prcomp object.

See also

Examples

m = matrix(rnorm(100),ncol=5)
extractPC(m)
#> Standard deviations (1, .., p=5):
#> [1] 2.657803e+00 2.304456e+00 2.128930e+00 1.960785e+00 6.095177e-16
#> 
#> Rotation (n x k) = (20 x 5):
#>               PC1          PC2          PC3         PC4          PC5
#>  [1,]  0.31525545 -0.271939921  0.225274774 -0.10942126  0.731587883
#>  [2,] -0.05681500 -0.191443052 -0.182132019 -0.19020980 -0.297425400
#>  [3,] -0.10858456  0.404838026  0.015210989  0.03393839  0.333197825
#>  [4,] -0.01887602 -0.211887994 -0.041854372 -0.07611815  0.292072456
#>  [5,]  0.03286103 -0.041286241  0.219548050 -0.12570910 -0.045513634
#>  [6,]  0.19292775  0.049661636  0.007833188  0.21337680 -0.008228419
#>  [7,] -0.41608958  0.289838579  0.097772186  0.05411148  0.123534507
#>  [8,] -0.08480583  0.007832205 -0.337836534 -0.11804556  0.136618856
#>  [9,] -0.05869767  0.023724388 -0.536928742  0.18702488  0.224518940
#> [10,]  0.21762108  0.071171008 -0.216848216  0.59298800  0.054204297
#> [11,] -0.26450625 -0.102904210 -0.281124848 -0.13240872  0.134294715
#> [12,] -0.19385762  0.269996639 -0.178748258 -0.21089737  0.090475149
#> [13,]  0.17755562  0.023160092 -0.042956062 -0.13229739 -0.041856973
#> [14,] -0.18129414 -0.174853786 -0.133564940 -0.06141440  0.051373149
#> [15,] -0.09665086  0.507496856  0.140667588  0.04775902  0.024232895
#> [16,] -0.43645953 -0.212924718  0.459529601  0.18704489  0.003527805
#> [17,]  0.08421784 -0.147548253 -0.168225094  0.11692608 -0.182180648
#> [18,] -0.25554432 -0.314686586  0.003342005  0.27132643 -0.048594156
#> [19,]  0.39851349  0.212286357  0.075469697 -0.22084739 -0.154303284
#> [20,]  0.11325414  0.058222225  0.105813522  0.47273565 -0.002500894