Build a two-column word/frequency table

Build a two-column word/frequency table

meshTable(
  RAVmodel,
  ind,
  rm.noise = NULL,
  weighted = TRUE,
  filterMessage = TRUE
)

Arguments

RAVmodel

A PCAGenomicSignatures object

ind

An index of RAV

rm.noise

An integer. Under the default (rm.noise=NULL), if cluster size (= s) is smaller than 8, rm.noise = floor(s*0.5). For clusters with >= 8 PCs, rm.noise = 4. If rm.noise = 0, all the MeSH terms in RAV will be used to draw wordcloud.

weighted

A logical. If TRUE, MeSH terms from each study are weighted based on the variance explained by the principle component of the study contributing a give RAV. Default is TRUE.

filterMessage

A logical. Under the default TRUE, any output RAV belong to the filtering list will give a message. Silence this message with filterMessage=FALSE. You can check the filter list using data("filterList").

Value

A table with two columns, word and freq. MeSH terms in the defined RAV (by ind argument) is ordered based on their frequency.

Examples

data(miniRAVmodel)
meshTable(miniRAVmodel,1139)
#> RAV1139 can be filtered based on GSEA_PLIERpriors
#>                                      word         freq
#> 21                  NOTCH1 protein, human 0.1242330284
#> 26                       Receptor, Notch1 0.1242330284
#> 16 Leukemia, Lymphocytic, Chronic, B-Cell 0.0745398170
#> 3                           CD28 Antigens 0.0656441334
#> 15                                 Infant 0.0468886667
#> 29                        T Cell Receptor 0.0468886667
#> 22                               Oncogene 0.0414110095
#> 18                  Lymphocyte Activation 0.0364689630
#> 14                                Indoles 0.0295148844
#> 1                           B-Lymphocytes 0.0266213632
#> 19                          Muromonab-CD3 0.0248496481
#> 9                         DNA Methylation 0.0113179540
#> 5                              Cell Cycle 0.0102568958
#> 4              CD4-Positive T-Lymphocytes 0.0091538548
#> 8                                   Child 0.0091172407
#> 20                               Mutation 0.0082822019
#> 33                Transcriptome Profiling 0.0071352319
#> 7                      Cell Proliferation 0.0065385804
#> 28                    Signal Transduction 0.0064258463
#> 11                    Epigenesis, Genetic 0.0063119359
#> 17               Leukemia, Myeloid, Acute 0.0056759393
#> 24                         RNA, Messenger 0.0044354144
#> 30                  Transcription Factors 0.0036068205
#> 6                               Cell Line 0.0032408616
#> 10                     Dimethyl Sulfoxide 0.0026352575
#> 12                                  Goals 0.0022703757
#> 34                 Whole Exome Sequencing 0.0018319490
#> 25                                RNA-Seq 0.0014356923
#> 31                          Transcriptome 0.0013462998
#> 32                 Transcriptome Analysis 0.0011295295
#> 13                                 Humans 0.0007936742
#> 23                                    RNA 0.0003967054
#> 2                           Base Sequence 0.0003965369
#> 35                                   mRNA 0.0003601398
#> 27                        Sequencing, RNA 0.0002691298