Subset enriched pathways of RAV

Subset enriched pathways of RAV

subsetEnrichedPathways(
  RAVmodel,
  ind = NULL,
  n = 10,
  both = FALSE,
  include_nes = FALSE
)

Arguments

RAVmodel

PCAGenomicSignatures object. Also an output from GSEA can be used.

ind

A numeric vector containing the RAV number you want to check enriched pathways. If not specified, this function returns results from all the RAVs.

n

The number of top and bottom pathways to be selected based on normalized enrichment score (NES).

both

Default is FALSE, where only the top n pathways will be printed. If it is set to TRUE, the output will contain both top and bottom n pathways.

include_nes

Defalt is FALSE. If it set to TRUE, the output will include both description and NES of the enriched pathway.

Value

A DataFrame with top and bottom n pathways from the enrichment results.

Examples

data(miniRAVmodel)

# all RAVS in model
subsetEnrichedPathways(miniRAVmodel,n=5)
#> DataFrame with 5 rows and 20 columns
#>         RAV1076.Description     RAV338.Description    RAV1467.Description
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_METABOLISM_.. MIPS_PA700_20S_PA28_.. REACTOME_MITOTIC_PRO..
#> Up_2 REACTOME_HOST_INTERA.. REACTOME_RESPIRATORY.. REACTOME_CHROMOSOME_..
#> Up_3       MIPS_SPLICEOSOME REACTOME_RESPIRATORY.. REACTOME_3_UTR_MEDIA..
#> Up_4       KEGG_SPLICEOSOME MIPS_55S_RIBOSOME_MI.. REACTOME_INFLUENZA_L..
#> Up_5 REACTOME_ASSEMBLY_OF.. KEGG_PARKINSONS_DISE..   REACTOME_TRANSLATION
#>         RAV1614.Description     RAV294.Description    RAV3071.Description
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_DNA_REPLICA.. REACTOME_CELL_CYCLE_.. KEGG_PURINE_METABOLISM
#> Up_2 REACTOME_MITOTIC_M_M.. REACTOME_M_G1_TRANSI.. MIPS_55S_RIBOSOME_MI..
#> Up_3 REACTOME_CELL_CYCLE_.. REACTOME_MITOTIC_PRO..       PID_FOXM1PATHWAY
#> Up_4    REACTOME_CELL_CYCLE MIPS_55S_RIBOSOME_MI.. REACTOME_E2F_MEDIATE..
#> Up_5 REACTOME_CELL_CYCLE_.. REACTOME_MRNA_PROCES..        PID_ATR_PATHWAY
#>         RAV1694.Description     RAV438.Description     RAV725.Description
#>                 <character>            <character>            <character>
#> Up_1 REACTOME_ADAPTIVE_IM.. REACTOME_METABOLISM_.. REACTOME_ADAPTIVE_IM..
#> Up_2 KEGG_PURINE_METABOLISM              DMAP_ERY5 REACTOME_METABOLISM_..
#> Up_3 REACTOME_SIGNALING_B..              DMAP_ERY4              DMAP_ERY5
#> Up_4 REACTOME_DOWNSTREAM_.. REACTOME_POST_TRANSL..              DMAP_ERY4
#> Up_5              DMAP_ERY4              DMAP_ERY3              DMAP_ERY3
#>         RAV1497.Description     RAV501.Description     RAV941.Description
#>                 <character>            <character>            <character>
#> Up_1 MIPS_28S_RIBOSOMAL_S.. REACTOME_TRANSCRIPTION IRIS_DendriticCell-C..
#> Up_2 MIPS_55S_RIBOSOME_MI.. REACTOME_METABOLISM_.. REACTOME_CHROMOSOME_..
#> Up_3 REACTOME_RESPIRATORY..    REACTOME_DNA_REPAIR       PID_PLK1_PATHWAY
#> Up_4 KEGG_OXIDATIVE_PHOSP.. REACTOME_RESPIRATORY.. IRIS_CD4Tcell-Th1-re..
#> Up_5 REACTOME_RESPIRATORY.. REACTOME_CYCLIN_E_AS.. REACTOME_M_G1_TRANSI..
#>         RAV2538.Description    RAV1139.Description     RAV884.Description
#>                 <character>            <character>            <character>
#> Up_1    REACTOME_CELL_CYCLE              DMAP_ERY3              DMAP_ERY3
#> Up_2 REACTOME_CELL_CYCLE_.. KEGG_ALZHEIMERS_DISE.. REACTOME_METABOLISM_..
#> Up_3 REACTOME_DNA_REPLICA.. REACTOME_POST_TRANSL.. KEGG_ALZHEIMERS_DISE..
#> Up_4 REACTOME_MITOTIC_M_M..     REACTOME_APOPTOSIS KEGG_HUNTINGTONS_DIS..
#> Up_5 REACTOME_MITOTIC_PRO.. KEGG_HUNTINGTONS_DIS..        KEGG_CELL_CYCLE
#>          RAV695.Description     RAV953.Description    RAV1994.Description
#>                 <character>            <character>            <character>
#> Up_1 IRIS_Bcell-Memory_Ig.. IRIS_Bcell-Memory_Ig.. SVM T cells CD4 memo..
#> Up_2           DMAP_BCELLA3  IRIS_Bcell-Memory_IgM           DMAP_TCELLA6
#> Up_3       IRIS_Bcell-naive       IRIS_Bcell-naive           DMAP_TCELLA8
#> Up_4  IRIS_Bcell-Memory_IgM           DMAP_BCELLA3           DMAP_TCELLA2
#> Up_5           DMAP_BCELLA2           DMAP_BCELLA1           DMAP_TCELLA7
#>          RAV312.Description     RAV468.Description
#>                 <character>            <character>
#> Up_1 REACTOME_CELL_CYCLE_.. REACTOME_METABOLISM_..
#> Up_2 REACTOME_MITOTIC_M_M..                     NA
#> Up_3 REACTOME_DNA_REPLICA..                     NA
#> Up_4    REACTOME_CELL_CYCLE                     NA
#> Up_5                     NA                     NA

# only a specific RAV (note the colnames above)
subsetEnrichedPathways(miniRAVmodel,ind=695,n=5)
#> DataFrame with 5 rows and 1 column
#>          RAV695.Description
#>                 <character>
#> Up_1 IRIS_Bcell-Memory_Ig..
#> Up_2           DMAP_BCELLA3
#> Up_3       IRIS_Bcell-naive
#> Up_4  IRIS_Bcell-Memory_IgM
#> Up_5           DMAP_BCELLA2